removed class CommandsList
