+++ .gitignore
@@ -1,2 +1,4 @@
 */bin
 */obj
+*.vs
+*/.vs
