+++ .vs/VSWorkspaceState.json
@@ -3,6 +3,6 @@
     "",
     "\\NailBot"
   ],
-  "SelectedNode": "\\NailBot\\Program.cs",
+  "SelectedNode": "\\NailBot",
   "PreviewInSolutionExplorer": false
 }
